+++ CHANGE_LOG.md
@@ -5,5 +5,8 @@
 
 ##V2.3.0
 
+### Major Changes
+* Massively update the failed downloads dialog
+
 ### Minor Changes
 * Fix title bar name not changing in last downloaded dialog